requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pyyaml>=6.0
open3d>=0.17.0        # optional: only needed for visualize mode
wandb>=0.16.0         # optional: experiment tracking
